[JENKINS-75258] Allow system configuration to be saved with plugin enabled
https://issues.jenkins.io/browse/JENKINS-75258 reports that the system
configuration cannot be saved from "Manage Jenkins" -> "System" when
the Conjur Credentials plugin is installed and enabled.  If the plugin
is disabled, the system configuration can be saved.  If the plugin is
not installed, the system configuration can be saved.

It appears that the disableSubmitButton JavaScript function is trying
to prevent the administrator from saving an incomplete or imperfect
configuration.  Administrators generally do not like configuration
forms that forbid saving an incomplete configuration.  They expect to
save part of the configuration and be able to return to complete the
configuration later.

The JavaScript debugging tools report an exception
attempting to assign the value true to a field `disabled`
of a null object `applyBtn`. That is line 192 of the file
src/main/resources/org/conjur/jenkins/configuration/GlobalConjurConfiguration/config.jelly

This commit removes the disableSubmitButton() call without removing the
implementation of the disableSubmitButton() method.  If the fundamental
issue can be fixed, then the call could be enabled again.